requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
colorama==0.4.3
Django==3.0.6
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dotenv==0.13.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
